carve up the monolith
