+++ opentargets_ontologyutils/rdf_utils.py
@@ -17,58 +17,6 @@ from datetime import date
 
 logger = logging.getLogger(__name__)
 
-EFO_TAS = [
-    'http://www.ebi.ac.uk/efo/EFO_1000018', # bladder disease
-    'http://www.ebi.ac.uk/efo/EFO_0000319', # cardiovascular disease
-    'http://www.ebi.ac.uk/efo/EFO_0000405', # digestive system disease
-    'http://www.ebi.ac.uk/efo/EFO_0001379', # endocrine
-    'http://www.ebi.ac.uk/efo/EFO_0003966', # eye disease
-    'http://www.ebi.ac.uk/efo/EFO_0000508', # genetic disorder
-    'http://www.ebi.ac.uk/efo/EFO_0000524', # head disease
-    'http://www.ebi.ac.uk/efo/EFO_0005803', # henmatological
-    'http://www.ebi.ac.uk/efo/EFO_0000540', # immune system disease
-    'http://www.ebi.ac.uk/efo/EFO_0003086', # kidney disease
-    'http://www.ebi.ac.uk/efo/EFO_0005741', # infection disease
-    'http://www.ebi.ac.uk/efo/EFO_0000589', # metabolic disease
-    'http://www.ebi.ac.uk/efo/EFO_0000616', # neoplasm
-    'http://www.ebi.ac.uk/efo/EFO_0000618', # nervous system
-    'http://www.ebi.ac.uk/efo/EFO_0000512', # reproductive system
-    'http://www.ebi.ac.uk/efo/EFO_0000684', # respiratory system
-    'http://www.ebi.ac.uk/efo/EFO_0002461', # skeletal system
-    'http://www.ebi.ac.uk/efo/EFO_0000701', # skin disease
-    'http://www.ebi.ac.uk/efo/EFO_0001421', # liver disease
-]
-
-HPO_TAS = [
-    "http://purl.obolibrary.org/obo/HP_0005386", #"behavior/neurological phenotype",
-    "http://purl.obolibrary.org/obo/HP_0005375", #"adipose tissue phenotype",
-    "http://purl.obolibrary.org/obo/HP_0005385", #"cardiovascular system phenotype",
-    "http://purl.obolibrary.org/obo/HP_0005384", #"cellular phenotype",
-    "http://purl.obolibrary.org/obo/HP_0005382", #"craniofacial phenotype",
-    "http://purl.obolibrary.org/obo/HP_0005381", #"digestive/alimentary phenotype",
-    "http://purl.obolibrary.org/obo/HP_0005380", #"embryo phenotype",
-    "http://purl.obolibrary.org/obo/HP_0005379", #"endocrine/exocrine phenotype",
-    "http://purl.obolibrary.org/obo/HP_0005378", #"growth/size/body region phenotype",
-    "http://purl.obolibrary.org/obo/HP_0005377", #"hearing/vestibular/ear phenotype",
-    "http://purl.obolibrary.org/obo/HP_0005397", #"hematopoietic system phenotype",
-    "http://purl.obolibrary.org/obo/HP_0005376", #"homeostasis/metabolism phenotype",
-    "http://purl.obolibrary.org/obo/HP_0005387", #"immune system phenotype",
-    "http://purl.obolibrary.org/obo/HP_0010771", #"integument phenotype",
-    "http://purl.obolibrary.org/obo/HP_0005371", #"limbs/digits/tail phenotype",
-    "http://purl.obolibrary.org/obo/HP_0005370", #"liver/biliary system phenotype",
-    "http://purl.obolibrary.org/obo/HP_0010768", #"mortality/aging",
-    "http://purl.obolibrary.org/obo/HP_0005369", #"muscle phenotype",
-    "http://purl.obolibrary.org/obo/HP_0002006", #"neoplasm",
-    "http://purl.obolibrary.org/obo/HP_0003631", #"nervous system phenotype",
-    "http://purl.obolibrary.org/obo/HP_0002873", #"normal phenotype",
-    "http://purl.obolibrary.org/obo/HP_0001186", #"pigmentation phenotype",
-    "http://purl.obolibrary.org/obo/HP_0005367", #"renal/urinary system phenotype",
-    "http://purl.obolibrary.org/obo/HP_0005389", #"reproductive system phenotype",
-    "http://purl.obolibrary.org/obo/HP_0005388", #"respiratory system phenotype",
-    "http://purl.obolibrary.org/obo/HP_0005390", #"skeleton phenotype",
-    "http://purl.obolibrary.org/obo/HP_0005394", #"taste/olfaction phenotype",
-    "http://purl.obolibrary.org/obo/HP_0005391", #"vision/eye phenotype"
-]
 
 '''
 subclass generators: yield a series of values
@@ -350,268 +298,6 @@ class OntologyClassReader(object):
                 rdf_properties[property_name] = [property_value]
         return rdf_properties
 
-    def load_hpo_classes(self, uri):
-        """Loads the HPO graph and extracts the current and obsolete classes.
-           Status: production
-        """
-        logger.debug("load_hpo_classes...")
-
-        self.load_ontology_graph(uri)
-
-        base_class = 'http://purl.obolibrary.org/obo/HP_0000118'
-        self.load_ontology_classes(base_class=base_class)
-        self.get_deprecated_classes()
-        self.get_top_levels(base_class= base_class)
-
-    def load_mp_classes(self, uri):
-        """Loads the HPO graph and extracts the current and obsolete classes.
-           Status: production
-        """
-        logger.debug("load_mp_classes...")
-
-        self.load_ontology_graph(uri)
-        base_class = 'http://purl.obolibrary.org/obo/MP_0000001'
-        self.load_ontology_classes(base_class= base_class)
-        self.get_deprecated_classes()
-        self.get_top_levels(base_class= base_class)
-
-        #self.get_ontology_top_levels(base_class, top_level_map=self.phenotype_top_levels)
-
-    def load_efo_classes(self, uri, lightweight=False, uberon_uri=None):
-        """Loads the EFO graph and extracts the current and obsolete classes.
-           Status: production
-        """
-        logger.debug("load_efo_classes...")
-
-        print("load EFO classes...")
-
-        self.load_ontology_graph(uri)
-
-        if uberon_uri:
-            logger.debug("load Uberon classes...")
-            print("load Uberon classes...")
-            self.load_ontology_graph(uberon_uri)
-
-        print("Done")
-        if lightweight == True:
-            return
-
-        # load disease, phenotype, measurement, biological process, function and injury and mental health
-        for base_class in [ 'http://www.ebi.ac.uk/efo/EFO_0000408',
-                            'http://www.ebi.ac.uk/efo/EFO_0000651',
-                            'http://www.ebi.ac.uk/efo/EFO_0001444',
-                            'http://purl.obolibrary.org/obo/GO_0008150',
-                            'http://www.ifomis.org/bfo/1.1/snap#Function',
-                            'http://www.ebi.ac.uk/efo/EFO_0000546',
-                            'http://www.ebi.ac.uk/efo/EFO_0003935' ]:
-            self.load_ontology_classes(base_class=base_class)
-            self.get_top_levels(base_class=base_class)
-        self.get_deprecated_classes()
-
-
-    #TODO this should be a static method
-    def get_hpo(self, uri):
-        '''
-        Load HPO to accept phenotype terms that are not in EFO
-        :return:
-        '''
-        obj = OntologyClassReader()
-        obj.load_hpo_classes(uri)
-        obj.rdf_graph = None
-        return obj
-
-    #TODO this should be a static method
-    def get_mp(self, uri):
-        '''
-        Load MP to accept phenotype terms that are not in EFO
-        :return:
-        '''
-        obj = None
-        obj = OntologyClassReader()
-        obj.load_mp_classes(uri)
-        obj.rdf_graph = None
-        return obj
-
-    #TODO this should be a static method
-    def get_efo(self, uri):
-        '''
-        Load EFO current and obsolete classes to report them to data providers
-        :return:
-        '''
-        obj = OntologyClassReader()
-        obj.load_efo_classes(uri)
-        obj.rdf_graph = None
-        return obj
-
-    def load_open_targets_disease_ontology(self, efo_uri):
-        """Loads the EFO graph and extracts the current and obsolete classes.
-           Generates the therapeutic areas
-           Creates the other disease class
-           Move injury under other disease
-           Status: production
-        """
-        logger.debug("load_open_targets_disease_ontology...")
-        self.load_ontology_graph(efo_uri)
-        all_ns = [n for n in self.rdf_graph.namespace_manager.namespaces()]
-
-        self.get_deprecated_classes(obsoleted_in_version=True)
-
-        '''
-        get the original top_levels from EFO
-        '''
-        self.get_top_levels(base_class='http://www.ebi.ac.uk/efo/EFO_0000408')
-
-        '''
-        Detach the TAs from the disease node
-        and load all the classes
-        '''
-        disease_uri = URIRef('http://www.ebi.ac.uk/efo/EFO_0000408')
-        for base_class in EFO_TAS:
-            uri = URIRef(base_class)
-            self.rdf_graph.remove((uri, None, disease_uri))
-            self.load_ontology_classes(base_class=base_class)
-            self.get_classes_paths(root_uri=base_class, level=0)
-
-        '''
-        Create an other disease node
-        '''
-        cttv = Namespace(str("http://www.targetvalidation.org/disease"))
-
-        # namespace_manager = NamespaceManager(self.rdf_graph)
-        self.rdf_graph.namespace_manager.bind('cttv', cttv)
-
-        '''
-        Some diseases have no categories, let's create a category for them
-        '''
-        other_disease_uri = URIRef('http://www.targetvalidation.org/disease/other')
-        self.rdf_graph.add((other_disease_uri, RDF.type, rdflib.term.URIRef(u'http://www.w3.org/2002/07/owl#Class')))
-        self.rdf_graph.add([other_disease_uri, RDFS.label, rdflib.Literal('other disease')])
-
-        '''
-        Get all children of 'disease' and assign them to 'other disease'
-        '''
-        for c in self.rdf_graph.subjects(predicate=RDFS.subClassOf, object=disease_uri):
-            self.rdf_graph.add([c, RDFS.subClassOf, other_disease_uri])
-
-        '''
-        Move 'injury' under 'other disease'
-        injuries are treated with medication recorded in ChEMBL
-        Move 'mental health' under 'other disease'
-        '''
-        self.rdf_graph.add([URIRef('http://www.ebi.ac.uk/efo/EFO_0000546'), RDFS.subClassOf, other_disease_uri])
-        self.rdf_graph.add([URIRef('http://www.ebi.ac.uk/efo/EFO_0003935'), RDFS.subClassOf, other_disease_uri])
-
-
-        # other disease, phenotype, measurement, biological process, function
-        for base_class in [ 'http://www.targetvalidation.org/disease/other',
-                            'http://www.ebi.ac.uk/efo/EFO_0000651',
-                            'http://www.ebi.ac.uk/efo/EFO_0001444',
-                            'http://purl.obolibrary.org/obo/GO_0008150',
-                            'http://www.ifomis.org/bfo/1.1/snap#Function']:
-
-            self.load_ontology_classes(base_class=base_class)
-            self.get_classes_paths(root_uri=base_class, level=0)
-
-    def load_mammalian_phenotype_ontology(self, uri):
-        """
-            Loads the MP graph and extracts the current and obsolete classes.
-            Status: production
-        """
-        logger.debug("load_mammalian_phenotype_ontology...")
-        self.load_ontology_graph(uri)
-
-        all_ns = [n for n in self.rdf_graph.namespace_manager.namespaces()]
-
-        '''
-        Detach the anatomical system from the mammalian phenotype node
-        and load all the classes
-        '''
-
-        mp_root_uri = 'http://purl.obolibrary.org/obo/MP_0000001'
-        mp_root_uriref = URIRef(mp_root_uri)
-        self.get_children(mp_root_uri)
-
-        for child in self.children[mp_root_uri]:
-            print("%s %s..."%(child['code'], child['label']))
-            uri = "http://purl.obolibrary.org/obo/" + child['code']
-            uriref = URIRef(uri)
-            self.rdf_graph.remove((uriref, None, mp_root_uriref))
-            self.load_ontology_classes(base_class=uri)
-            self.get_classes_paths(root_uri=uri, level=0)
-            print(len(self.current_classes))
-
-    def load_evidence_classes(self, uri_so, uri_eco):
-        '''
-        Loads evidence from ECO, SO and the Open Targets evidence classes
-        :return:
-        '''
-        self.load_ontology_graph(uri_so)
-        self.load_ontology_graph(uri_eco)
-
-        evidence_uri = URIRef('http://purl.obolibrary.org/obo/ECO_0000000')
-
-        # for triple in self.rdf_graph.triples((evidence_uri, None, None)):
-        #      logger.debug(triple)
-
-        '''
-            Open Targets specific evidence:
-            A) Create namespace for OT evidence
-            B) Add Open Targets evidence terms to graph
-            C) Traverse the graph breadth first
-        '''
-
-        cttv = Namespace(str("http://www.targetvalidation.org/disease"))
-        ot = Namespace(str("http://identifiers.org/eco"))
-
-        #namespace_manager = NamespaceManager(self.rdf_graph)
-        self.rdf_graph.namespace_manager.bind('cttv', cttv)
-        self.rdf_graph.namespace_manager.bind('ot',ot)
-
-        open_targets_terms = {
-            'http://www.targetvalidation.org/disease/cttv_evidence':'CTTV evidence',
-            'http://identifiers.org/eco/drug_disease':'drug-disease evidence',
-            'http://identifiers.org/eco/target_drug':'biological target to drug evidence',
-            'http://identifiers.org/eco/clinvar_gene_assignments':'ClinVAR SNP-gene pipeline',
-            'http://identifiers.org/eco/cttv_mapping_pipeline':'CTTV-custom annotation pipeline',
-            'http://identifiers.org/eco/GWAS':'Genome-wide association study evidence',
-            'http://identifiers.org/eco/GWAS_fine_mapping': 'Fine-mapping study evidence',
-            'http://identifiers.org/eco/somatic_mutation_evidence':'Somatic mutation evidence',
-            'http://www.targetvalidation.org/evidence/genomics_evidence':'genomics evidence',
-            'http://targetvalidation.org/sequence/nearest_gene_five_prime_end':'Nearest gene counting from 5&apos; end',
-            'http://targetvalidation.org/sequence/regulatory_nearest_gene_five_prime_end':'Nearest regulatory gene from 5&apos; end',
-            'http://www.targetvalidation.org/evidence/literature_mining':'Literature mining',
-            'http://www.targetvalidation.org/provenance/DatabaseProvenance':'database provenance',
-            'http://www.targetvalidation.org/provenance/ExpertProvenance':'expert provenance',
-            'http://www.targetvalidation.org/provenance/GWAS_SNP_to_trait_association':'GWAS SNP to trait association',
-            'http://www.targetvalidation.org/provenance/LiteratureProvenance':'literature provenance',
-            'http://www.targetvalidation.org/provenance/disease_to_phenotype_association':'disease to phenotype association',
-            'http://www.targetvalidation.org/provenance/gene_to_disease_association':'gene to disease association'
-        }
-
-        for uri, label in open_targets_terms.items():
-            u = URIRef(uri)
-            self.rdf_graph.add((u, RDF.type, rdflib.term.URIRef(u'http://www.w3.org/2002/07/owl#Class')))
-            self.rdf_graph.add([u, RDFS.label, rdflib.Literal(label)])
-            self.rdf_graph.add([u, RDFS.subClassOf, evidence_uri])
-
-        u = URIRef('http://identifiers.org/eco/target_drug')
-        # for triple in self.rdf_graph.triples((u, None, None)):
-        #      logger.debug(triple)
-
-        #(a, b, c) = self.rdf_graph.namespace_manager.compute_qname(unicode('http://identifiers.org/eco/target_drug'))
-        #logger.debug(c)
-        #return
-        # Add the bits specific to Open Targets
-        # 'Open Targets evidence' ?
-        # 'biological target-disease association via drug' ECO:0000360
-        # 'drug-disease evidence' http://identifiers.org/eco/drug_disease SubclassOf 'biological target-disease association via drug'
-        # 'biological target to drug evidence' http://identifiers.org/eco/target_drug SubclassOf 'biological target-disease association via drug'
-        # ClinVAR SNP-gene pipeline http://identifiers.org/eco/clinvar_gene_assignments SubclassOf ECO:0000246
-        # CTTV-custom annotation pipeline http://identifiers.org/eco/cttv_mapping_pipeline SubclassOf ECO:0000246
-
-        for base_class in ['http://purl.obolibrary.org/obo/ECO_0000000', 'http://purl.obolibrary.org/obo/SO_0000400', 'http://purl.obolibrary.org/obo/SO_0001260', 'http://purl.obolibrary.org/obo/SO_0000110', 'http://purl.obolibrary.org/obo/SO_0001060' ]:
-            self.load_ontology_classes(base_class=base_class)
-            self.get_classes_paths(root_uri=base_class, level=0)
 
 class DiseaseUtils(object):
